Inflow fix

+++ app/src/main/java/com/example/android/inventorymanager/Models/TransactionDetail.java
@@ -8,7 +8,7 @@ import java.util.Date;
 
 public class TransactionDetail {
     public String user;
-    public int amount;
+    public long amount;
     public long timestamp;
     //public String quantity;
